requirements.txt
pandas>=2.2
numpy>=1.26
yfinance>=0.2.54
PyYAML>=6.0.2
